[worker] Split files (#12623)
